+++ src/main/java/tukorea_2024_s3_10/eat_fit/application/README.md
@@ -1,0 +1,1 @@
+# application
