+++ src/main/java/tukorea_2024_s3_10/eat_fit/domain/user/entity/User.java
@@ -1,0 +1,34 @@
+package tukorea_2024_s3_10.eat_fit.domain.user.entity;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.hibernate.annotations.CreationTimestamp;
+
+import java.sql.Timestamp;
+
+@Entity
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class User {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private Long id;
+
+    private String oauthId;
+
+    @Enumerated(EnumType.STRING)
+    private Role role;
+
+    @CreationTimestamp
+    private Timestamp createdAt;
+
+    public void changeRole(Role role){
+        this.role = role;
+    }
+}
