+++ src/main/java/tukorea_2024_s3_10/eat_fit/domain/user/repository/UserProfileRepository.java
@@ -1,0 +1,7 @@
+package tukorea_2024_s3_10.eat_fit.domain.user.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import tukorea_2024_s3_10.eat_fit.domain.user.entity.UserProfile;
+
+public interface UserProfileRepository extends JpaRepository<UserProfile, Long> {
+}
