+++ src/main/java/tukorea_2024_s3_10/eat_fit/global/README.md
@@ -1,0 +1,8 @@
+# global
+프로젝트 전역으로 적용되는 파일들이 존재한다.
+
+ex) ApiResponse 객체는 모든 응답에 일관되게 사용된다.
+
+ex) WebConfig 프로젝트 전역 레벨로 설정되는 파일이다.
+
+ex) GlobalExceptionHandler 전역적으로 발생하는 예외를 잡는 핸들러다.
