+++ src/main/java/tukorea_2024_s3_10/eat_fit/global/exception/GlobalExceptionHandler.java
@@ -1,0 +1,15 @@
+package tukorea_2024_s3_10.eat_fit.global.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+import tukorea_2024_s3_10.eat_fit.global.dto.ApiResponse;
+
+import java.util.Map;
+
+@RestControllerAdvice
+public class GlobalExceptionHandler {
+
+
+}
