--- src/main/java/tukorea_2024_s3_10/eat_fit/global/exception/handler/GlobalExceptionHandler.java
@@ -1,20 +1,0 @@
-package tukorea_2024_s3_10.eat_fit.global.exception.handler;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-import tukorea_2024_s3_10.eat_fit.global.dto.ApiResponse;
-import tukorea_2024_s3_10.eat_fit.user.exception.EmailAlreadyExistsException;
-
-import java.util.Map;
-
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity<ApiResponse<Map<String, String>>> handleEmailAlreadyExists(EmailAlreadyExistsException e) {
-        return ResponseEntity.status(HttpStatus.CONFLICT)
-                .body(ApiResponse.error(e.getMessage()));
-    }
-}
