+++ src/main/java/tukorea_2024_s3_10/eat_fit/infrastructure/README.md
@@ -1,0 +1,5 @@
+# infrastructure
+
+외부 기술 및 라이브러리에 의존되는 코드들이 존재하는 곳
+
+ex) SQL 쿼리를 다루는 Repository 구현체
