--- src/main/java/tukorea_2024_s3_10/eat_fit/infrastructure/security/CustomSuccessHandler.java
@@ -1,59 +1,0 @@
-package tukorea_2024_s3_10.eat_fit.infrastructure.security;
-
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.Cookie;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.web.authentication.SimpleUrlAuthenticationSuccessHandler;
-import org.springframework.stereotype.Component;
-import tukorea_2024_s3_10.eat_fit.infrastructure.security.jwt.JwtUtil;
-import tukorea_2024_s3_10.eat_fit.infrastructure.security.dto.CustomOAuth2User;
-
-import java.io.IOException;
-import java.util.Collection;
-import java.util.Iterator;
-
-@Component
-@RequiredArgsConstructor
-public class CustomSuccessHandler extends SimpleUrlAuthenticationSuccessHandler {
-
-    private final JwtUtil jwtUtil;
-
-    @Override
-    public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException, ServletException {
-
-        //OAuth2User
-        CustomOAuth2User customUserDetails = (CustomOAuth2User) authentication.getPrincipal();
-
-        String username = customUserDetails.getUsername();
-
-        Collection<? extends GrantedAuthority> authorities = authentication.getAuthorities();
-        Iterator<? extends GrantedAuthority> iterator = authorities.iterator();
-        GrantedAuthority auth = iterator.next();
-        String role = auth.getAuthority();
-
-        String token = jwtUtil.createJwt(username, role, 60*60*60L*1000);
-
-        response.addCookie(createCookie("Authorization", token));
-
-        if(role.equals("ROLE_GUEST")) {
-            response.sendRedirect("http://localhost:3000/physical-info");
-            return;
-        }
-        response.sendRedirect("http://localhost:3000");
-    }
-
-    private Cookie createCookie(String key, String value) {
-
-        Cookie cookie = new Cookie(key, value);
-        cookie.setMaxAge(60*60*60);
-        //cookie.setSecure(true);
-        cookie.setPath("/");
-        cookie.setHttpOnly(true);
-
-        return cookie;
-    }
-}
