--- src/main/java/tukorea_2024_s3_10/eat_fit/infrastructure/security/dto/UserDto.java
@@ -1,11 +1,0 @@
-package tukorea_2024_s3_10.eat_fit.infrastructure.security.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class UserDto {
-    String role;
-    String username;
-}
