+++ src/main/java/tukorea_2024_s3_10/eat_fit/presentation/README.md
@@ -1,0 +1,7 @@
+# presentation
+
+클라이언트의 요청과 응답을 처리하는 계층이다.
+
+우리의 프로젝트에서는 Controller가 대표적인 예시이다.
+
+클라이언트의 요청을 먼저 받는 계층이기에 Request DTO는 여기서 관리한다.
