--- src/main/java/tukorea_2024_s3_10/eat_fit/user/exception/EmailAlreadyExistsException.java
@@ -1,7 +1,0 @@
-package tukorea_2024_s3_10.eat_fit.user.exception;
-
-public class EmailAlreadyExistsException extends RuntimeException {
-    public EmailAlreadyExistsException(String message) {
-        super(message);
-    }
-}
