--- src/main/java/tukorea_2024_s3_10/eat_fit/user/exception/handler/UserExceptionHandler.java
@@ -1,20 +1,0 @@
-package tukorea_2024_s3_10.eat_fit.user.exception.handler;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-import tukorea_2024_s3_10.eat_fit.global.dto.ApiResponse;
-import tukorea_2024_s3_10.eat_fit.user.exception.EmailAlreadyExistsException;
-
-import java.util.Map;
-
-@RestControllerAdvice(basePackages = "tukorea_2024_s3_10.eat_fit.user")
-public class UserExceptionHandler {
-
-    @ExceptionHandler(EmailAlreadyExistsException.class)
-    public ResponseEntity<ApiResponse<Map<String, String>>> handleEmailAlreadyExists(EmailAlreadyExistsException ex) {
-        return ResponseEntity.status(HttpStatus.CONFLICT)
-                .body(ApiResponse.fail(Map.of("email", ex.getMessage())));
-    }
-}
