--- src/main/java/tukorea_2024_s3_10/eat_fit/user/repository/UserGoalRepository.java
@@ -1,7 +1,0 @@
-package tukorea_2024_s3_10.eat_fit.user.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import tukorea_2024_s3_10.eat_fit.user.entity.UserGoal;
-
-public interface UserGoalRepository extends JpaRepository<UserGoal, Long> {
-}
